ao_refactor: move share into a component
